requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pymongo==4.8.0
boto3==1.35.20
python-dotenv==1.0.1
pandas==2.2.3
watchfiles